pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "thesis-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dropbox>=12.0.2",
    "pymupdf>=1.23.0",
    "python-pptx>=1.0.2",
    "openai>=1.0.0",
    "neo4j>=5.0.0",
    "python-dotenv>=1.0.0",
    "networkx>=3.0",
    "matplotlib>=3.7.0",
    "plotly>=5.17.0",
    "pandas>=2.0.0",
    "jupyter>=1.0.0",
    "numpy>=1.24.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
